converted the polls module to use the action controllers
