+++ ACP3/Modules/ACP3/Polls/Controller/Admin/Index/Index.php
@@ -1,0 +1,78 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\Polls\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Polls;
+
+/**
+ * Class Index
+ * @package ACP3\Modules\ACP3\Polls\Controller\Admin\Index
+ */
+class Index extends Core\Modules\AdminController
+{
+    /**
+     * @var \ACP3\Modules\ACP3\Polls\Model\PollRepository
+     */
+    protected $pollRepository;
+
+    /**
+     * Index constructor.
+     *
+     * @param \ACP3\Core\Modules\Controller\AdminContext    $context
+     * @param \ACP3\Modules\ACP3\Polls\Model\PollRepository $pollRepository
+     */
+    public function __construct(
+        Core\Modules\Controller\AdminContext $context,
+        Polls\Model\PollRepository $pollRepository
+    )
+    {
+        parent::__construct($context);
+
+        $this->pollRepository = $pollRepository;
+    }
+
+    /**
+     * @return array
+     */
+    public function execute()
+    {
+        $polls = $this->pollRepository->getAllInAcp();
+
+        /** @var Core\Helpers\DataGrid $dataGrid */
+        $dataGrid = $this->get('core.helpers.data_grid');
+        $dataGrid
+            ->setResults($polls)
+            ->setRecordsPerPage($this->user->getEntriesPerPage())
+            ->setIdentifier('#acp-table')
+            ->setResourcePathDelete('admin/polls/index/delete')
+            ->setResourcePathEdit('admin/polls/index/edit');
+
+        $dataGrid
+            ->addColumn([
+                'label' => $this->translator->t('system', 'publication_period'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\DateColumnRenderer::class,
+                'fields' => ['start', 'end'],
+                'default_sort' => true
+            ], 30)
+            ->addColumn([
+                'label' => $this->translator->t('polls', 'question'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\TextColumnRenderer::class,
+                'fields' => ['title'],
+            ], 20)
+            ->addColumn([
+                'label' => $this->translator->t('system', 'id'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\IntegerColumnRenderer::class,
+                'fields' => ['id'],
+                'primary' => true
+            ], 10);
+
+        return [
+            'grid' => $dataGrid->render(),
+            'show_mass_delete_button' => count($polls) > 0
+        ];
+    }
+}
