+++ ACP3/Modules/ACP3/Polls/Controller/Sidebar/Index/Index.php
@@ -1,13 +1,16 @@
 <?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
 
-namespace ACP3\Modules\ACP3\Polls\Controller\Sidebar;
+namespace ACP3\Modules\ACP3\Polls\Controller\Sidebar\Index;
 
 use ACP3\Core;
 use ACP3\Modules\ACP3\Polls;
 
 /**
  * Class Index
- * @package ACP3\Modules\ACP3\Polls\Controller\Sidebar
+ * @package ACP3\Modules\ACP3\Polls\Controller\Sidebar\Index
  */
 class Index extends Core\Modules\Controller
 {
@@ -50,7 +53,7 @@ class Index extends Core\Modules\Controller
         $this->voteRepository = $voteRepository;
     }
 
-    public function actionIndex()
+    public function execute()
     {
         $poll = $this->pollRepository->getLatestPoll($this->date->getCurrentDateTime());
